Extract V2 parts of Job and Build serializers
Job and Build serializers were cluttered with both V2 and V3 code, so I
decided to extract V2 code into separate files in order to make it
easier to remove it once we can do it.
