+++ app/serializers/build.js
@@ -1,48 +1,11 @@
-import V2FallbackSerializer from 'travis/serializers/v2_fallback';
+import BuildV2FallbackSerializer from 'travis/serializers/build_v2_fallback';
 
-var Serializer = V2FallbackSerializer.extend({
+var Serializer = BuildV2FallbackSerializer.extend({
   attrs: {
     _config: { key: 'config' },
     _finishedAt: { key: 'finished_at' },
     _startedAt: { key: 'started_at' },
     _duration: { key: 'duration' }
-  },
-
-  extractRelationships: function (modelClass, resourceHash) {
-    return this._super(modelClass, resourceHash);
-  },
-
-  normalizeSingleResponse: function (store, primaryModelClass, payload/* , id, requestType*/) {
-    if (payload.commit) {
-      payload.build.commit = payload.commit;
-      delete payload.build.commit_id;
-    }
-    return this._super(...arguments);
-  },
-
-  normalizeArrayResponse: function (store, primaryModelClass, payload/* , id, requestType*/) {
-    if (payload.commits) {
-      payload.builds.forEach(function (build) {
-        let commit = payload.commits.findBy('id', build.commit_id);
-        if (commit) {
-          build.commit = commit;
-          return delete build.commit_id;
-        }
-      });
-    }
-    return this._super(...arguments);
-  },
-
-  keyForV2Relationship: function (key/* , typeClass, method*/) {
-    if (key === 'jobs') {
-      return 'job_ids';
-    } else if (key === 'repo') {
-      return 'repository_id';
-    } else if (key === 'commit') {
-      return key;
-    } else {
-      return this._super(...arguments);
-    }
   },
 
   keyForRelationship(key/* , typeClass, method*/) {
@@ -54,54 +17,11 @@ var Serializer = V2FallbackSerializer.extend({
   },
 
   normalize: function (modelClass, resourceHash) {
-    // TODO: remove this after switching to V3 entirely
-    let type = resourceHash['@type'];
-    let commit = resourceHash.commit;
     if (resourceHash['event_type'] == 'pull_request' &&
           !resourceHash.hasOwnProperty('pull_request')) {
       // in V3 we don't return "pull_request" property as we rely on event_type
       // value. This line makes V3 payloads also populate pull_request property
       resourceHash['pull_request'] = true;
-    }
-    // when we get the commit from V3 we need to put a branch on a commit as the
-    // current commit model expects it
-    if (type && commit && resourceHash.branch && !commit.branch) {
-      let name = resourceHash.branch.name;
-      if (!name) {
-        let match = resourceHash.branch['@href'].match(/\/repo\/.*?\/branch\/(.*)/);
-        if (match) {
-          name = match[1];
-        }
-      }
-      if (name) {
-        commit.branch = name;
-      }
-    }
-    if (!type && commit && commit.hasOwnProperty('branch_is_default')) {
-      let build = resourceHash.build,
-        commit = resourceHash.commit;
-      let branch = {
-        name: commit.branch,
-        default_branch: commit.branch_is_default,
-        '@href': `/repo/${build.repository_id}/branch/${commit.branch}`
-      };
-      resourceHash.build.branch = branch;
-    }
-
-    // fix pusher payload, it doesn't include a branch record:
-    if (!resourceHash['@type'] && resourceHash.build &&
-       resourceHash.repository && resourceHash.repository.default_branch) {
-      let branchName = resourceHash.build.branch,
-        repository = resourceHash.repository,
-        defaultBranchName = repository.default_branch.name;
-
-      resourceHash.build.branch = {
-        name: branchName,
-        default_branch: branchName === defaultBranchName,
-        '@href': `/repo/${repository.id}/branch/${branchName}`
-      };
-
-      repository.default_branch['@href'] = `/repo/${repository.id}/branch/${defaultBranchName}`;
     }
 
     return this._super(modelClass, resourceHash);
